starting to extract methods to separate files
